Restructure a generic import dialog preparing for the addition of more imports
